Added calculator functionality
Implemented a way for the dimensions of the shapes to be calculated and
then displayed in the relevant textboxes.
Code for Square is complete.
Code for Circle is incomplete.
Code for triangle is incomplete.
I am currently working on implementing a method that will set the
correct values in the textboxes using only this method.