Major changes #2
Changed slider so that it doesn't multiply it's dimension by 2 anymore.
Changed setTextBox method parameter to accept ints. This should be
changed back to accepting a string.
Updated shape dimension variable so now there are only three(one for
each shape). Deleted redundant methods as a result.
Need to implement the new classes in the form class.